pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-dir]
"" = "."

[project]
name = "alki"
version = "0.1.0"
description = "Alki 🌊 — an open-source toolchain for deploying LLMs at the edge"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
dependencies = [
  "typer[all]>=0.9",
  "transformers>=4.55",
  "huggingface_hub>=0.25",
  "llama-cpp-python>=0.3.16",
  "safetensors>=0.4",
  "pyyaml>=6.0",
  "rich>=14.0",
  "psutil>=7.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "black",
  "ruff",
]

[project.scripts]
alki = "src.cli.main:app"
